Updates to Get Interface Closer
Changed the Card class to extend JLabel to make it easier to add them
to the CardPanel. Updated Player so that the Hit method returns the
Card as well as adding it to the hand so that I can add it to the
CardPanel. Updated the CardPanel to comment out the add of the cards,
doing this dynamically for now. Added a bunch of logic to the main
class to figure out the winner and such.