Updated user.java
Created getters for accessing number of (games) wins, lost and the logic
to track money won/lost.  Also changed the way the filename is created
by using the players name and adding the extension .gambit.  The
filename is saved in all lower case.  Also added a method to create a
new user with defaults if no file is found.